// frontend/src/components/post/template/carouselEditor/canvas/types.ts
// Type definitions for Canvas component
import { Media } from "../../../MediaUploader";

// Add TextAlign type
export type TextAlign = "left" | "center" | "right" | "justify";

// Define interface for Canvas ref
export interface CanvasRef {
  captureCanvasContent: (boxIndex: number) => Promise<Blob | null>;
  getTotalBoxes: () => number;
}

export interface ImageItem {
  id: string;
  src: string;
  position: { x: number; y: number };
  size: { width: number; height: number };
  zIndex: number;
}

export interface TextStyle {
  fontSize: number;
  color: string;
  fontFamily?: string;
  fontWeight?: string;
  fontStyle?: string;
  lineHeight?: string;
  textAlign?: TextAlign;
  rotation?: number;
}

export interface TextItem {
  id: string;
  content: string;
  position: { x: number; y: number };
  size?: { width: number | string; height: number | string };
  style: TextStyle;
  zIndex: number;
}

export interface CanvasItem {
  id: string;
  position: { x: number; y: number };
  zIndex: number;
  type: "image" | "text";
}

export type CanvasImageItem = ImageItem & { type: "image" };
export type CanvasTextItem = TextItem & { type: "text" };

export interface ImageToEdit {
  id: string;
  src: string;
}

export interface ImageToReplace {
  id: string;
  position: { x: number; y: number };
  size: { width: number; height: number };
}